+++ src/test/scala/com/github/plume/oss/drivers/OverflowDbTests.scala
@@ -1,13 +1,18 @@
 package com.github.plume.oss.drivers
 
 import com.github.plume.oss.testfixtures.PlumeDriverFixture
+import com.github.plume.oss.testfixtures.PlumeDriverFixture.{b1, m1}
+import io.shiftleft.codepropertygraph.generated.{Cpg, EdgeTypes}
+import io.shiftleft.passes.IntervalKeyPool
+import org.apache.tinkerpop.gremlin.tinkergraph.structure.TinkerGraph
+import overflowdb.BatchedUpdate
 
 import java.io.{File => JFile}
 import java.lang.reflect.Field
 import java.nio.charset.StandardCharsets
 import java.nio.file.{Files, Paths}
 import scala.io.Source
-import scala.util.Try
+import scala.util.{Failure, Success, Try}
 
 class OverflowDbTests extends PlumeDriverFixture(new OverflowDbDriver()) {
 
@@ -38,6 +43,32 @@ class OverflowDbTests extends PlumeDriverFixture(new OverflowDbDriver()) {
           Some(Source.fromInputStream(Files.newInputStream(methodSemanticsPath)))
         )
     }
+  }
+
+  "should be able to serialize and deserialize XML graphs without throwing an exception" in {
+    createSimpleGraph(driver)
+    val td      = driver.asInstanceOf[OverflowDbDriver]
+    val outFile = Paths.get("./odbGraph.xml").toFile
+    td.exportAsGraphML(outFile)
+    // Should be valid if TinkerGraph can accept it
+    Try({
+      val graph = TinkerGraph.open()
+      graph.traversal().io[Any](outFile.getAbsolutePath).read().iterate()
+      graph.close()
+    }) match {
+      case Failure(e) => fail("TinkerGraph could not import ODB generated XML", e)
+      case _          =>
+    }
+    outFile.delete()
+  }
+
+  private def createSimpleGraph(driver: IDriver): Unit = {
+    val cpg       = Cpg.empty
+    val keyPool   = new IntervalKeyPool(1, 1000)
+    val diffGraph = new BatchedUpdate.DiffGraphBuilder()
+    diffGraph.addNode(m1).addNode(b1).addEdge(m1, b1, EdgeTypes.AST)
+    val adg = BatchedUpdate.applyDiff(cpg.graph, diffGraph, keyPool, null)
+    driver.bulkTx(adg)
   }
 
 }