+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- `OverflowDbDriver::exportAsGraphML` for writing the graph instance as an XML file.
+
 ## [1.1.5] - 2022-03-18
 
 ### Changed